pyproject.toml
[project]
name = "tapayoka-pi"
version = "0.1.0"
description = "Tapayoka BLE peripheral firmware for Raspberry Pi"
requires-python = ">=3.11"
dependencies = [
    "bluezero>=0.8.0",
    "eth-account>=0.13.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
    "mypy>=1.10",
]
gpio = [
    "RPi.GPIO>=0.7.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
